Implement controllers

+++ src/main/resources/db/migration/V1.1__schema.sql
@@ -2,7 +2,7 @@ create sequence global_sequence increment by 10;
 
 create table people (
   id      integer     not null constraint people_pkey primary key,
-  version integer,
+  version smallint,
   name    varchar(32) not null,
   email   varchar(32) not null constraint uk_people_email unique,
   address jsonb       not null
@@ -10,7 +10,7 @@ create table people (
 
 create table cars (
   id        integer     not null constraint cars_pkey primary key,
-  version   integer,
+  version   smallint,
   brand     varchar(16) not null,
   model     varchar(16) not null,
   plate_num varchar(16) not null constraint uk_cars_plate_num unique,
